% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{AddEmptyRowsAtTop}
\alias{AddEmptyRowsAtTop}
\title{AddEmptyRowsAtTop
This adds empty rows to the top of a data.frame}
\usage{
AddEmptyRowsAtTop(d, numRows)
}
\arguments{
\item{d}{The data.frame of interest}

\item{numRows}{The number of rows to add on the top}
}
\description{
AddEmptyRowsAtTop
This adds empty rows to the top of a data.frame
}
