% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialise.R
\name{AllowFileManipulationFromInitialiseProject}
\alias{AllowFileManipulationFromInitialiseProject}
\title{Allows for InitialiseProject to manipulate files
on your system}
\usage{
AllowFileManipulationFromInitialiseProject()
}
\description{
Allows for InitialiseProject to manipulate files
on your system
}
