% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censor.R
\name{Censor}
\alias{Censor}
\title{Provides censoring of data}
\usage{
Censor(n, randomNoise, d = NULL, boundaries = NULL)
}
\arguments{
\item{n}{a number}

\item{randomNoise}{a number}

\item{d}{a number}

\item{boundaries}{a number}
}
\description{
Provides censoring of data
}
