% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_environments.r
\docType{data}
\name{CONFIG}
\alias{CONFIG}
\title{Environment for holding configs}
\format{An object of class \code{environment} of length 1.}
\usage{
CONFIG
}
\description{
Environment for holding configs
}
\keyword{datasets}
