% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_environments.r
\docType{data}
\name{CONFIG_STACK}
\alias{CONFIG_STACK}
\title{Environment for holding configs relevant to stack analyses}
\format{An object of class \code{environment} of length 1.}
\usage{
CONFIG_STACK
}
\description{
Environment for holding configs relevant to stack analyses
}
\keyword{datasets}
