% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cran.r
\name{CreateLocalCRAN}
\alias{CreateLocalCRAN}
\title{Creates own personal local CRAN
Taken from https://rstudio.github.io/packrat/custom-repos.html}
\usage{
CreateLocalCRAN(localCRAN = "H:/CRAN", libs = c("data.table", "tidyr",
  "magrittr", "grid", "stringr", "R6", "ggplot2", "dplyr", "gridExtra", "knitr",
  "rmarkdown", "brew", "scales", "Hmisc", "Gmisc", "htmlTable", "ggvis",
  "Formula", "roxygen2", "packrat", "rms", "survey"), overwrite = FALSE)
}
\description{
Creates own personal local CRAN
Taken from https://rstudio.github.io/packrat/custom-repos.html
}

