% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R
\name{DashboardEmail}
\alias{DashboardEmail}
\title{Sends out mass emails that are stored in an xlsx file}
\usage{
DashboardEmail(emailBCC, emailSubject, emailText,
  XLSXLocation = file.path("etc", "gmailr", "emails.xlsx"),
  OAUTHLocation = file.path("etc", "gmailr", ".httr-oauth"))
}
\arguments{
\item{emailBCC}{a}

\item{emailSubject}{a}

\item{emailText}{a}

\item{XLSXLocation}{a}

\item{OAUTHLocation}{a}
}
\description{
Sends out mass emails that are stored in an xlsx file
}
