% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R
\name{DashboardEmailSpecific}
\alias{DashboardEmailSpecific}
\title{Sends out mass emails that are stored in an xlsx file}
\usage{
DashboardEmailSpecific(emailBCC, emailSubject, emailText,
  OAUTHLocation = file.path("/etc", "gmailr", ".httr-oauth"))
}
\arguments{
\item{emailBCC}{a}

\item{emailSubject}{a}

\item{emailText}{a}

\item{OAUTHLocation}{a}
}
\description{
Sends out mass emails that are stored in an xlsx file
}
