% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R
\name{DashboardFolder}
\alias{DashboardFolder}
\title{If folders are setup according to the
dashboard philosophy, then this function
finds folders according to the dashboard}
\usage{
DashboardFolder(inside = "data_raw", f = NULL)
}
\arguments{
\item{inside}{where it is inside}

\item{f}{an optional file}
}
\description{
If folders are setup according to the
dashboard philosophy, then this function
finds folders according to the dashboard
}
