% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R
\name{DashboardInitialiseClean}
\alias{DashboardInitialiseClean}
\title{If folders are setup according to the
dashboard philosophy, then this function
sets RPROJ}
\usage{
DashboardInitialiseClean()
}
\description{
If folders are setup according to the
dashboard philosophy, then this function
sets RPROJ
}
