% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractExposureConfounders.R
\name{DetectSpline}
\alias{DetectSpline}
\title{Detect spline
Transforms cos and sin to amplitude peak and trough}
\usage{
DetectSpline(var)
}
\arguments{
\item{var}{The results}
}
\description{
Detect spline
Transforms cos and sin to amplitude peak and trough
}
