% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackAnalyses.R
\name{ExpandStack.int}
\alias{ExpandStack.int}
\title{ExpandStack.int
This validates the skeleton analysis stack}
\usage{
ExpandStack.int(stack, i = 1, newAnalysisID = FALSE)
}
\arguments{
\item{stack}{stack}

\item{i}{The variable of interest}

\item{newAnalysisID}{If you want new analysis IDs generated (i.e. link broken between expanded stack and old stack)}
}
\description{
ExpandStack.int
This validates the skeleton analysis stack
}
