% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackAnalyses.R
\name{ExpandStack}
\alias{ExpandStack}
\title{ExpandStack
This validates the skeleton analysis stack}
\usage{
ExpandStack(stack, newAnalysisID = FALSE)
}
\arguments{
\item{stack}{stack}

\item{newAnalysisID}{If you want new analysis IDs generated (i.e. link broken between expanded stack and old stack)}
}
\description{
ExpandStack
This validates the skeleton analysis stack
}
