% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractExposureConfounders.R
\name{ExtractExposureConfounders.int}
\alias{ExtractExposureConfounders.int}
\title{ExtractExposureConfounders.int
Transforms cos and sin to amplitude peak and trough}
\usage{
ExtractExposureConfounders.int(var)
}
\arguments{
\item{var}{The results}
}
\description{
ExtractExposureConfounders.int
Transforms cos and sin to amplitude peak and trough
}
