% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackAnalyses.R
\name{ExtractFits}
\alias{ExtractFits}
\title{ExtractFits
Extract fits}
\usage{
ExtractFits(fit0, fit1, fit1aic)
}
\arguments{
\item{fit0}{The variable of interest}

\item{fit1}{The variable of interest}

\item{fit1aic}{with a possibly more friendly family (possion instead of quasipoisson)}
}
\description{
ExtractFits
Extract fits
}
