% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackAnalyses.R
\name{ExtractFitsSplinesInteractions}
\alias{ExtractFitsSplinesInteractions}
\title{ExtractFitsSplines
Extract fits}
\usage{
ExtractFitsSplinesInteractions(fit0, fit1, fit1aic, stack, i, data, form,
  exposureValue = 1)
}
\arguments{
\item{fit0}{The variable of interest}

\item{fit1}{The variable of interest}

\item{fit1aic}{with a possibly more friendly family (possion instead of quasipoisson)}

\item{stack}{stack}

\item{i}{i}

\item{data}{data}

\item{form}{the formula}

\item{exposureValue}{exposureValue}
}
\description{
ExtractFitsSplines
Extract fits
}
