% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileStable.R, R/TempDir.R
\name{FileTimeModified}
\alias{FileTimeModified}
\alias{FileTimeModified}
\title{FileTimeModified
This determines if the file has been changed (checks modification times)}
\usage{
FileTimeModified(file = tempfile())

FileTimeModified(file = tempfile())
}
\arguments{
\item{file}{The file of interest}

\item{file}{The file of interest}
}
\description{
FileTimeModified
This determines if the file has been changed (checks modification times)

FileTimeModified
This determines if the file has been changed (checks modification times)
}
