% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.r
\name{FootnoteGridArrange}
\alias{FootnoteGridArrange}
\title{Makes a footnote on the bottom right hand corner}
\usage{
FootnoteGridArrange(footnoteText = strftime(Sys.time(), format =
  "\%d/\%m/\%Y"), size = 1.2, color = "black")
}
\arguments{
\item{footnoteText}{a}

\item{size}{a}

\item{color}{a}
}
\description{
Makes a footnote on the bottom right hand corner
}
