% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackAnalyses.R
\name{FormatPValue}
\alias{FormatPValue}
\title{FormatPValue
This formats the results from the stack}
\usage{
FormatPValue(pval)
}
\arguments{
\item{pval}{The results}
}
\description{
FormatPValue
This formats the results from the stack
}
