% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackAnalyses.R
\name{FormatResultsStack}
\alias{FormatResultsStack}
\title{FormatResultsStack
This formats the results from the stack}
\usage{
FormatResultsStack(results, bonf)
}
\arguments{
\item{results}{The results}

\item{bonf}{do we want bonferroni correction?}
}
\description{
FormatResultsStack
This formats the results from the stack
}
