% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.r
\name{FortifyShapeFileNorwegianCounties}
\alias{FortifyShapeFileNorwegianCounties}
\title{Produces norwegian counties fortified file}
\usage{
FortifyShapeFileNorwegianCounties()
}
\description{
Produces norwegian counties fortified file
}
