% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.r
\name{FortifyShapeFileNorwegianMunicipalities}
\alias{FortifyShapeFileNorwegianMunicipalities}
\title{Produces norwegian municipalities fortified file}
\usage{
FortifyShapeFileNorwegianMunicipalities()
}
\description{
Produces norwegian municipalities fortified file
}
