% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/knitr.R
\name{GenHTMLReport}
\alias{GenHTMLReport}
\title{Uses preset CSS to knit html report
Also defaults to base64 inline images}
\usage{
GenHTMLReport(inFile = "", outFile = "")
}
\description{
Uses preset CSS to knit html report
Also defaults to base64 inline images
}

