% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggplot.r
\name{ggplotFormat}
\alias{ggplotFormat}
\title{Formats ggplot to a more minimalistic style}
\usage{
ggplotFormat(q, sizeMultiplier = 1, legendKey = 1, xAngle = 0,
  stripes = FALSE)
}
\description{
Formats ggplot to a more minimalistic style
}

