% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialise.R
\name{Initialise}
\alias{Initialise}
\title{If folders are setup according to the
dashboard philosophy, then this function
sets RPROJ}
\usage{
Initialise()
}
\description{
If folders are setup according to the
dashboard philosophy, then this function
sets RPROJ
}
