% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialise.R
\name{InitialiseOpinionatedUnix}
\alias{InitialiseOpinionatedUnix}
\title{Initialises project on a unix machine in an opinionated manner}
\usage{
InitialiseOpinionatedUnix(project)
}
\arguments{
\item{project}{location of project}
}
\description{
Initialises project on a unix machine in an opinionated manner
}
