% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialise.R
\name{InitialiseProject}
\alias{InitialiseProject}
\title{Initialises project}
\usage{
InitialiseProject(PROJHOME = NULL, PROJRAW = NULL, PROJCLEAN = NULL,
  PROJBAKED = NULL, PROJFINAL = NULL, PROJSHARED = NULL, HOME = NULL,
  RAW = NULL, CLEAN = NULL, BAKED = NULL, FINAL = NULL, SHARED = NULL)
}
\arguments{
\item{PROJHOME}{a}

\item{PROJRAW}{a}

\item{PROJCLEAN}{a}

\item{PROJBAKED}{a}

\item{PROJFINAL}{a}

\item{PROJSHARED}{a}

\item{HOME}{a}

\item{RAW}{a}

\item{CLEAN}{a}

\item{BAKED}{a}

\item{FINAL}{a}

\item{SHARED}{a}
}
\description{
Initialises project
}
