% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileStable.R
\name{IsFileChanged}
\alias{IsFileChanged}
\title{IsFileChanged
This determines if the file has been changed (checks modification times)}
\usage{
IsFileChanged(fileToCheck = tempfile(), fileDetails = tempfile(),
  delay = 5)
}
\arguments{
\item{fileToCheck}{The file of interest}

\item{fileDetails}{File that contains the details of things being checked}

\item{delay}{Number of seconds between checking file size a second time}
}
\description{
IsFileChanged
This determines if the file has been changed (checks modification times)
}
