% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileStable.R
\name{IsFileStable}
\alias{IsFileStable}
\title{Is the file stable (size not increasing/decreasing)?}
\usage{
IsFileStable(file = tempfile(), delay = 5)
}
\arguments{
\item{file}{The file of interest}

\item{delay}{Number of seconds between checking file size a second time}
}
\description{
Is the file stable (size not increasing/decreasing)?
}
