% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractExposureConfounders.R
\name{IsInteraction}
\alias{IsInteraction}
\title{Is the variable here an interaction?}
\usage{
IsInteraction(var)
}
\arguments{
\item{var}{The results}
}
\description{
Is the variable here an interaction?
}
