% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/libraries.r
\name{Libraries}
\alias{Libraries}
\title{Bulk loads the most common analysis libraries}
\usage{
Libraries(lib = c("data.table", "tidyr", "magrittr", "grid", "stringr", "R6",
  "ggplot2", "dplyr", "gridExtra", "knitr", "knitcitations", "bibtex",
  "rmarkdown", "brew", "scales", "Hmisc", "Gmisc", "htmlTable", "ggvis",
  "Formula", "roxygen2", "packrat", "rms", "survey"))
}
\description{
Bulk loads the most common analysis libraries
}

