% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackAnalyses.R
\name{LRTest}
\alias{LRTest}
\title{LRTest
Likelihood ratio test wrapper}
\usage{
LRTest(fit0, fit1)
}
\arguments{
\item{fit0}{The variable of interest}

\item{fit1}{The variable of interest}
}
\description{
LRTest
Likelihood ratio test wrapper
}
