% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggplot.r
\name{MakeFootnote}
\alias{MakeFootnote}
\title{Makes a footnote on the bottom right hand corner}
\usage{
MakeFootnote(footnoteText = strftime(Sys.time(), format = "\%d/\%m/\%Y"),
  size = 1.2, color = "black")
}
\description{
Makes a footnote on the bottom right hand corner
}

