% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PkgdownUndocumented.R
\name{PkgdownUndocumented}
\alias{PkgdownUndocumented}
\title{PkgdownUndocumented}
\usage{
PkgdownUndocumented(pkgroot = ".", pkgdown = file.path(pkgroot, "pkgdown",
  "_pkgdown.yml"))
}
\arguments{
\item{pkgroot}{Root the package}

\item{pkgdown}{Location of the pkgdown.yml file}
}
\description{
PkgdownUndocumented
}
