% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggplot.r
\name{png_a4}
\alias{png_a4}
\title{Creates png with default dimensions of A4}
\usage{
png_a4(file = "Figure.png", w = 1, h = 1, landscape = FALSE)
}
\description{
Creates png with default dimensions of A4
}

