% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/docx_document.r
\name{prCaptionFix}
\alias{prCaptionFix}
\title{Fixes the caption back fro elements withou}
\usage{
prCaptionFix(outFile)
}
\arguments{
\item{outFile}{The name of the file}
}
\value{
outFile The name of the file
}
\description{
Fixes the caption back fro elements withou
}
\keyword{internal}

