% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/docx_document.r
\name{prFtpOtherChanges}
\alias{prFtpOtherChanges}
\title{Changes lines for XML-conformity}
\usage{
prFtpOtherChanges(output_str)
}
\arguments{
\item{output_str}{The input from readLines()}
}
\value{
string Returns with changes
}
\description{
Changes lines for XML-conformity
}
\keyword{internal}

