% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackAnalyses.R
\name{ProcessStack}
\alias{ProcessStack}
\title{ProcessStack
This processes one value from the stack}
\usage{
ProcessStack(stack, i = 1)
}
\arguments{
\item{stack}{The stack}

\item{i}{The i'th stack value}
}
\description{
ProcessStack
This processes one value from the stack
}
