% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialise.R
\name{RCloneSync}
\alias{RCloneSync}
\title{Wrapper for RClone Sync}
\usage{
RCloneSync(from, to, createLocalIfRemoteDoesntExist = FALSE)
}
\arguments{
\item{from}{a}

\item{to}{a}

\item{createLocalIfRemoteDoesntExist}{a}
}
\description{
Wrapper for RClone Sync
}
