% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatable.R
\name{RecodeDT}
\alias{RecodeDT}
\title{Allows for recoding of variables}
\usage{
RecodeDT(d, switch, var)
}
\arguments{
\item{d}{a}

\item{switch}{a vector coded like: c("old"="new")}

\item{var}{a}
}
\description{
Allows for recoding of variables
}
