% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr.R
\name{RmdToDOCX}
\alias{RmdToDOCX}
\title{Wrapper around rmarkdown::render for docx files}
\usage{
RmdToDOCX(inFile = "", outFile = "", copyFrom = NULL, ...)
}
\arguments{
\item{inFile}{a}

\item{outFile}{a}

\item{copyFrom}{a}

\item{...}{a}
}
\description{
Wrapper around rmarkdown::render for docx files
}
