% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr.R
\name{RmdToHTML}
\alias{RmdToHTML}
\title{Wrapper around rmarkdown::render for HTML files}
\usage{
RmdToHTML(inFile = "", outFile = "", copyFrom = NULL, ...)
}
\arguments{
\item{inFile}{a}

\item{outFile}{a}

\item{copyFrom}{a}

\item{...}{a}
}
\description{
Wrapper around rmarkdown::render for HTML files
}
