% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.r
\name{SMAOpng}
\alias{SMAOpng}
\title{Creates png with default dimensions of half A4}
\usage{
SMAOpng(file = "Figure.png", w = 1, h = 1, landscape = TRUE)
}
\arguments{
\item{file}{a}

\item{w}{a}

\item{h}{a}

\item{landscape}{a}
}
\description{
Creates png with default dimensions of half A4
}
