% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{SummarizeContinuous}
\alias{SummarizeContinuous}
\title{SummarizeContinuous
This summarizes continuous variables}
\usage{
SummarizeContinuous(var)
}
\arguments{
\item{var}{The variable of interest}
}
\description{
SummarizeContinuous
This summarizes continuous variables
}
