% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{SummarizeDispatch}
\alias{SummarizeDispatch}
\title{SummarizeDispatch
Summarizes one variable, with the option of stratifying by a 0/1 variable}
\usage{
SummarizeDispatch(var, by = NULL, labelLeft = NULL, labelTop = NULL,
  labelTop0 = NULL, labelTop1 = NULL)
}
\arguments{
\item{var}{The variable of interest}

\item{by}{A 0/1 variable to be stratified by (optional)}

\item{labelLeft}{Adding a label to the left of the returned data.frame}

\item{labelTop}{Adding a label to the top of the returned data.frame (only to be used if by=NULL)}

\item{labelTop0}{Adding a label to the top left of the returned data.frame (only to be used if by!=NULL)}

\item{labelTop1}{Adding a label to the top right of the returned data.frame (only to be used if by!=NULL)}
}
\description{
SummarizeDispatch
Summarizes one variable, with the option of stratifying by a 0/1 variable
}
