% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{SummarizeDispatchOneVar}
\alias{SummarizeDispatchOneVar}
\title{SummarizeDispatchOneVar
Summarizes one variable (mostly for internal use)}
\usage{
SummarizeDispatchOneVar(var, labelLeft = NULL, labelTop = NULL)
}
\arguments{
\item{var}{The variable of interest}

\item{labelLeft}{Adding a label to the left of the returned data.frame}

\item{labelTop}{Adding a label to the top of the returned data.frame}
}
\description{
SummarizeDispatchOneVar
Summarizes one variable (mostly for internal use)
}
