% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.r
\name{theme_SMAO}
\alias{theme_SMAO}
\title{Wrapper for ggplot theme}
\usage{
theme_SMAO(base_size = 24, base_family = "", v = 1)
}
\arguments{
\item{base_size}{base size}

\item{base_family}{text family}

\item{v}{version}
}
\description{
Wrapper for ggplot theme
}
