% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.r
\name{theme_SMAO45}
\alias{theme_SMAO45}
\title{Internal version 1 of a minimalist ggplot theme}
\usage{
theme_SMAO45(base_size = 12, base_family = "")
}
\arguments{
\item{base_size}{base size}

\item{base_family}{text family}
}
\description{
Internal version 1 of a minimalist ggplot theme
}
