% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackAnalyses.R
\name{TransformCosSinToAmplitudePeakTrough}
\alias{TransformCosSinToAmplitudePeakTrough}
\title{TransformCosSinToAmplitudePeakTrough
Transforms cos and sin to amplitude peak and trough}
\usage{
TransformCosSinToAmplitudePeakTrough(cos_b, sin_b)
}
\arguments{
\item{cos_b}{The results}

\item{sin_b}{The results}
}
\description{
TransformCosSinToAmplitudePeakTrough
Transforms cos and sin to amplitude peak and trough
}
