% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialise.R
\name{UseRClone}
\alias{UseRClone}
\title{Allows for InitialiseProject to manipulate files
on your system}
\usage{
UseRClone()
}
\description{
Allows for InitialiseProject to manipulate files
on your system
}
