% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackAnalyses.R
\name{ExpandStack}
\alias{ExpandStack}
\title{ExpandStack
This validates the skeleton analysis stack}
\usage{
ExpandStack(stack)
}
\arguments{
\item{stack}{stack}
}
\description{
ExpandStack
This validates the skeleton analysis stack
}
