% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackAnalyses.R
\name{ExtractFitsSplines}
\alias{ExtractFitsSplines}
\title{ExtractFitsSplines
Extract fits}
\usage{
ExtractFitsSplines(fit0, fit1, fit1aic, stack, i, data)
}
\arguments{
\item{fit0}{The variable of interest}

\item{fit1}{The variable of interest}

\item{fit1aic}{with a possibly more friendly family (possion instead of quasipoisson)}

\item{stack}{stack}

\item{i}{i}

\item{data}{data}
}
\description{
ExtractFitsSplines
Extract fits
}
