% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialise.R
\name{InitialiseProject}
\alias{InitialiseProject}
\title{Initialises project}
\usage{
InitialiseProject(HOME = NULL, RAW = NULL, CLEAN = NULL, BAKED = NULL,
  FINAL = NULL, SHARED = NULL)
}
\arguments{
\item{HOME}{a}

\item{RAW}{a}

\item{CLEAN}{a}

\item{BAKED}{a}

\item{FINAL}{a}

\item{SHARED}{a}
}
\description{
Initialises project
}
