% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialise.R
\name{InitialiseProject}
\alias{InitialiseProject}
\title{Initialises project}
\usage{
InitialiseProject(HOME = NULL, RAW = NULL, CLEAN = NULL, BAKED = NULL,
  FINAL = NULL, SHARED = NULL, RCLONE_RAW = NULL, RCLONE_CLEAN = NULL,
  RCLONE_BAKED = NULL, RCLONE_FINAL = NULL, RCLONE_SHARED = NULL)
}
\arguments{
\item{HOME}{a}

\item{RAW}{a}

\item{CLEAN}{a}

\item{BAKED}{a}

\item{FINAL}{a}

\item{SHARED}{a}

\item{RCLONE_RAW}{a}

\item{RCLONE_CLEAN}{a}

\item{RCLONE_BAKED}{a}

\item{RCLONE_FINAL}{a}

\item{RCLONE_SHARED}{a}
}
\description{
Initialises project
}
