% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/knitr.R
\name{RmdToDOCX}
\alias{RmdToDOCX}
\title{Uses preset CSS to knit html report
Also defaults to base64 inline images
CSS file taken from Max Gordon (http://gforge.se/packages/)
If pandoc is available, it uses pandoc (and hence bibliography/citations)
Otherwise uses knitr and no bibliography/citations}
\usage{
RmdToDOCX(inFile = "", outFile = "", tocDepth = 2)
}
\description{
Uses preset CSS to knit html report
Also defaults to base64 inline images
CSS file taken from Max Gordon (http://gforge.se/packages/)
If pandoc is available, it uses pandoc (and hence bibliography/citations)
Otherwise uses knitr and no bibliography/citations
}

