% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/knitr.R
\name{RmdToHTML}
\alias{RmdToHTML}
\title{Uses preset CSS to knit html report
Also defaults to base64 inline images}
\usage{
RmdToHTML(inFile = "", outFile = "")
}
\description{
Uses preset CSS to knit html report
Also defaults to base64 inline images
}

